Move `SkinnableElementTargetContainer` out of HUD namespace
